refactor: put common test code in separate file, export certain utilities
